backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# MongoDB
pymongo
motor

# Embeddings
sentence-transformers

# PDF Parsing
PyMuPDF

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
numpy

# Logging
loguru

# File Uploads
aiofiles

# Process
tqdm